requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
google-cloud-bigquery==3.33.0